+++ assets/i18n/en.json
@@ -200,12 +200,12 @@
     "I18N_LEARNER_DASHBOARD_COMPLETED_SECTION": "Completed",
     "I18N_LEARNER_DASHBOARD_COMPLETED_TO_INCOMPLETE_COLLECTIONS": "<[numberMoved]> of the collections you completed have been moved to the 'in progress' section as new explorations have been added to them!",
     "I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_COLLECTION": "<[numberDeleted]> of the collections which you completed has been recently deleted. We are sorry for the inconvenience",
-    "I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATION": "<[numberDeleted]> of the explorations which you completed has been recently deleted. We are sorry for the inconvenience",
-    "I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTION": "<[numberDeleted]> of the collections in progress has been recently deleted. We are sorry for the inconvenience",
-    "I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATION": "<[numberDeleted]> of the explorations in progress has been recently deleted. We are sorry for the inconvenience",
     "I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_COLLECTIONS": "<[numberDeleted]> of the collections which you completed have been recently deleted. We are sorry for the inconvenience",
+    "I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATION": "<[numberDeleted]> of the explorations which you completed has been recently deleted. We are sorry for the inconvenience",
     "I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATIONS": "<[numberDeleted]> of the explorations which you completed have been recently deleted. We are sorry for the inconvenience",
+    "I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTION": "<[numberDeleted]> of the collections in progress has been recently deleted. We are sorry for the inconvenience",
     "I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTIONS": "<[numberDeleted]> of the collections in progress have been recently deleted. We are sorry for the inconvenience",
+    "I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATION": "<[numberDeleted]> of the explorations in progress has been recently deleted. We are sorry for the inconvenience",
     "I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATIONS": "<[numberDeleted]> of the explorations in progress have been recently deleted. We are sorry for the inconvenience",
     "I18N_LEARNER_DASHBOARD_EMPTY_COMPLETED_COLLECTIONS": "It looks like you haven't completed any collections yet. Head over to the library to start an exciting new collection!",
     "I18N_LEARNER_DASHBOARD_EMPTY_COMPLETED_EXPLORATIONS": "It looks like you haven't completed any explorations yet. Head over to the library to start an exciting new exploration!",
