+++ assets/i18n/hi.json
@@ -200,12 +200,12 @@
 	"I18N_LEARNER_DASHBOARD_COMPLETED_SECTION": "पूर्ण",
 	"I18N_LEARNER_DASHBOARD_COMPLETED_TO_INCOMPLETE_COLLECTIONS": "आपके द्वारा किए गए संग्रह में से <[numberMoved]> को 'प्रगति' अनुभाग में स्थानांतरित किया गया क्योंकि उन्हें नए अन्वेषण जोड़े गए हैं!",
 	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_COLLECTION": "आपके द्वारा पूरा किए गए संग्रह में से <[numberDeleted]> को हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
-	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATION": "आपके द्वारा पूरी की गई अन्वेषण में से <[numberDeleted] को हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
-	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTION": "<[numberDeleted]> संग्रह की प्रगति में हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
-	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATION": "<[numberDeleted]> प्रगति में अन्वेषण हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
 	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_COLLECTIONS": "आपके द्वारा पूरा किए गए संग्रहों में से <[numberDeleted]> हाल ही में हटाए गए हैं असुविधा के लिए खेद है",
+	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATION": "आपके द्वारा पूरी की गई अन्वेषण में से <[numberDeleted] को हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
 	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATIONS": "आपके द्वारा पूरा किए गए अन्वेषण के <[numberDeleted]> हाल ही में हटाए गए हैं असुविधा के लिए खेद है",
+	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTION": "<[numberDeleted]> संग्रह की प्रगति में हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
 	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTIONS": "प्रगति में <[numberDeleted]> संग्रह हाल ही में हटाए गए हैं असुविधा के लिए खेद है",
+	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATION": "<[numberDeleted]> प्रगति में अन्वेषण हाल ही में हटा दिया गया है असुविधा के लिए खेद है",
 	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATIONS": "<[numberDeleted]> प्रगति में अन्वेषण हाल ही में हटाए गए हैं असुविधा के लिए खेद है",
 	"I18N_LEARNER_DASHBOARD_EMPTY_COMPLETED_COLLECTIONS": "ऐसा लगता है कि आपने अभी तक कोई भी कलेक्शन्स पूरा नहीं किया है। एक नया रोमांचक कलेक्शन्स शुरू करने के लिए पुस्तकालय में जाएं!",
 	"I18N_LEARNER_DASHBOARD_EMPTY_COMPLETED_EXPLORATIONS": "ऐसा लगता है कि आपने अभी तक कोई अन्वेषण पूरा नहीं किया है। एक रोमांचक नई अन्वेषण शुरू करने के लिए पुस्तकालय पर जाएं!",
