+++ assets/i18n/qqq.json
@@ -200,12 +200,12 @@
 	"I18N_LEARNER_DASHBOARD_COMPLETED_SECTION": "Text for the completed section in the learner dashboard.",
 	"I18N_LEARNER_DASHBOARD_COMPLETED_TO_INCOMPLETE_COLLECTIONS": "Text displayed in the learner dashboard. - The text is shown when new exporations are added to the collections the user has already completed.",
 	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_COLLECTION": "Text displayed in the learner dashboard. - This text is shown when a completed collection has been deleted.",
-	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATION": "Text displayed in the learner dashboard. - This text is shown when a completed exploration has been deleted.",
-	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTION": "Text displayed in the learner dashboard. - This text is shown when a collection in progress has been deleted.",
-	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATION": "Text displayed in the learner dashboard. - This text is shown when an exploration in progress has been deleted.",
 	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_COLLECTIONS": "Text displayed in the learner dashboard. - This text is shown when more than one completed collection has been deleted.",
+	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATION": "Text displayed in the learner dashboard. - This text is shown when a completed exploration has been deleted.",
 	"I18N_LEARNER_DASHBOARD_DELETED_COMPLETED_EXPLORATIONS": "Text displayed in the learner dashboard. - This text is shown when more than one completed exploration has been deleted.",
+	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTION": "Text displayed in the learner dashboard. - This text is shown when a collection in progress has been deleted.",
 	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_COLLECTIONS": "Text displayed in the learner dashboard. - This text is shown when more than one collection in progress has been deleted.",
+	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATION": "Text displayed in the learner dashboard. - This text is shown when an exploration in progress has been deleted.",
 	"I18N_LEARNER_DASHBOARD_DELETED_INCOMPLETE_EXPLORATIONS": "Text displayed in the learner dashboard. - This text is shown when more than one exploration in progress has been deleted.",
 	"I18N_LEARNER_DASHBOARD_EMPTY_COMPLETED_COLLECTIONS": "Text displayed in the learner dashboard. - This text is shown when there are no collections in the completed section of the user.",
 	"I18N_LEARNER_DASHBOARD_EMPTY_COMPLETED_EXPLORATIONS": "Text displayed in the learner dashboard. - This text is shown when there are no explorations in the completed section of the user.",
