+++ core/controllers/learner_dashboard_test.py
@@ -1,0 +1,125 @@
+# Copyright 2014 The Oppia Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for the learner dashboard and the notifications dashboard."""
+
+from core.domain import learner_progress_services
+from core.domain import subscription_services
+from core.tests import test_utils
+import feconf
+
+class LearnerDashboardHandlerTest(test_utils.GenericTestBase):
+
+    OWNER_EMAIL = 'owner@example.com'
+    OWNER_USERNAME = 'owner'
+
+    EXP_ID = 'exp_id'
+    EXP_TITLE = 'Exploration title'
+
+    COL_ID = 'col_id'
+    COL_TITLE = 'Collection title'
+
+    def setUp(self):
+        super(LearnerDashboardHandlerTest, self).setUp()
+        self.signup(self.OWNER_EMAIL, self.OWNER_USERNAME)
+        self.signup(self.VIEWER_EMAIL, self.VIEWER_USERNAME)
+
+        self.viewer_id = self.get_user_id_from_email(self.VIEWER_EMAIL)
+        self.owner_id = self.get_user_id_from_email(self.OWNER_EMAIL)
+
+    def test_can_see_completed_explorations(self):
+        self.login(self.VIEWER_EMAIL)
+
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['completed_explorations_list']), 0)
+
+        self.save_new_default_exploration(
+            self.EXP_ID, self.owner_id, title=self.EXP_TITLE)
+
+        learner_progress_services.mark_exploration_as_completed(
+            self.viewer_id, self.EXP_ID)
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['completed_explorations_list']), 1)
+        self.assertEqual(
+            response['completed_explorations_list'][0]['id'], self.EXP_ID)
+        self.logout()
+
+    def test_can_see_completed_collections(self):
+        self.login(self.VIEWER_EMAIL)
+
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['completed_collections_list']), 0)
+
+        self.save_new_default_collection(
+            self.COL_ID, self.owner_id, title=self.COL_TITLE)
+
+        learner_progress_services.mark_collection_as_completed(
+            self.viewer_id, self.COL_ID)
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['completed_collections_list']), 1)
+        self.assertEqual(
+            response['completed_collections_list'][0]['id'], self.COL_ID)
+        self.logout()
+
+    def test_can_see_incomplete_explorations(self):
+        self.login(self.VIEWER_EMAIL)
+
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['incomplete_explorations_list']), 0)
+
+        self.save_new_default_exploration(
+            self.EXP_ID, self.owner_id, title=self.EXP_TITLE)
+
+        state_name = 'state_name'
+        version = 1
+
+        learner_progress_services.mark_exploration_as_incomplete(
+            self.viewer_id, self.EXP_ID, state_name, version)
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['incomplete_explorations_list']), 1)
+        self.assertEqual(
+            response['incomplete_explorations_list'][0]['id'], self.EXP_ID)
+        self.logout()
+
+    def test_can_see_incomplete_collections(self):
+        self.login(self.VIEWER_EMAIL)
+
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['incomplete_collections_list']), 0)
+
+        self.save_new_default_collection(
+            self.COL_ID, self.owner_id, title=self.COL_TITLE)
+
+        learner_progress_services.mark_collection_as_incomplete(
+            self.viewer_id, self.COL_ID)
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['incomplete_collections_list']), 1)
+        self.assertEqual(
+            response['incomplete_collections_list'][0]['id'], self.COL_ID)
+        self.logout()
+
+    def test_can_see_subscription(self):
+        self.login(self.VIEWER_EMAIL)
+
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['subscription_list']), 0)
+
+        subscription_services.subscribe_to_creator(
+            self.viewer_id, self.owner_id)
+        response = self.get_json(feconf.LEARNER_DASHBOARD_DATA_URL)
+        self.assertEqual(len(response['subscription_list']), 1)
+        self.assertEqual(
+            response['subscription_list'][0]['creator_username'],
+            self.OWNER_USERNAME)
+        self.logout()
